Re-configure Jupyterhub to be better usable by a shared team.
Traefik is added as a reverse proxy.  It is configured for a
particular domain such as example.com.  That domain needs to have a
TLS cetificate which can be generated either manually by Let's Encrypt
or by Traefik itself.  It also needs to refreshed from time to
time (should happen automatically).

Traefik listens to request on port 443 only.  It only accepts HTTPS
and not HTTP.  When the request comes to example.com:443/jupyter then
it is routed to the "hub" container which is running the JupyterHub.
JupyterHub itself needs to be adjusted to know that it's base URL now
includes /jupyter which is done in c.JupyterHub.base_url.

The users sign up on Jupyterhub and are given access to their own
environment that is defined by the separate container that JupyterHub
spins up for them called jupterhubnotebook.  This means that the
container image needs to be built before hand.  That is done by the
command in the playbook:

docker build -f Dockerfile.jupyternotebook . -t jupyternotebook --ssh
default && docker compose up --build

The ssh arguments for building the container are such that they mount
the existing SSH credentials of the host for the duration of the
build.  This is needed to be able to git clone a repo, including
private ones.  It is important to pass uid=$NB_UID even though it is
not documented anywhere.

The host needs to have the SSH agent running and the required
identities need to be added with ssh-add so that they show up in
ssh-add -l.  This is because what is actually mounted during the build
is $SSH_AUTH_SOCK which is just a way for the container to access the
in-memory ssh private keys that are held by the ssh-agent.  It is
important that all components are accessible to the jovyan user rather
than just root.

The environment also creates a shared folder, a private folder and a
data folder.  A shared folder is mapped to the same volume so that
users share their notebooks simply.  This also needs to be configured
in c.DockerSpawner.volumes so that Jupyerhub mounts the needed volumes
on the go.  It doesn't work to mount a volume inside a directory tree
of another volume, so instead 3 directories trees are created for them
and then symbolic links are created in the main working area to
simulate the nesting.

Getting a GPU working inside that per-user container is a whole other
can of works.  It is critical that  c.DockerSpawner passses runtime
and device_requests parameters which is an undocumented way equivalent
to passing --gpus all as normally.

The docker daemon needs to be reconfigured to have default runtime be
nvidia so that it is the runtime used during the build because there
is no way to pass a runtime argument to the build command.

The docker daemon for some reason also needs to manually be supplied
Google's DNS addresses otherwise individual users aren't able to
access internet from within their containers.

GRANT_SUDO and --user=root is needed so that users are allowed
passwordless sudo inside their container.

The host needs to have nvidia container toolkit installed.  The Nvidia
drivers need to be accompannied with a supported version of CUDA.  It
appears in 2024 one doesn't have to worry about manually installing
CUDA Toolkit, cuDNN and their version anymore.